Split the Configuration file, as it was getting too large
