+++ kliff/dataset/__init__.py
@@ -1,4 +1,5 @@
-from .dataset import Configuration, Dataset
+from .configuration import Configuration
+from .dataset import Dataset
 from .extxyz import read_extxyz, write_extxyz
 
 __all__ = ["Configuration", "Dataset", "read_extxyz", "write_extxyz"]
